Final lab 12 - web

+++ SpringMVCTesting/src/main/java/com/tecsup/SpringMVC/dao/EmployeeDAO.java
@@ -15,7 +15,7 @@ public interface EmployeeDAO {
 
 	void delete(String login) throws DAOException;
 
-	void update(String login, String password, String lastname, String firstname, int salary, int dptId) throws DAOException;
+	void update(String login, String password, String lastname, String firstname, long salary, int dptId) throws DAOException;
 
 	Employee findEmployeeByLogin(String login) throws DAOException, EmptyResultException;
